// frontend/src/api/schema.gen.ts
export interface components {
    schemas: {
// ...
        DemandOut: {
            /**
             * Date
             * Format: date
             */
            date: string;
            /** Start Min */
            start_min: number;
            /** End Min */
            end_min: number;
            /** Headcount */
            headcount: number;
            /** Skill Value Id */
            skill_value_id?: number | null;
            /** Id */
            id: number;
        };
// ...
        ProfileOut: {
            /** Name */
            name: string;
            /**
             * Config
             * @default {}
             */
            config: Record<string, never>;
            /** Id */
            id: number;
        };
// ...
        RosterCreate: {
            /** Profile Id */
            profile_id: number;
            /** Name */
            name: string;
            /**
             * Roster Start
             * Format: date
             */
            roster_start: string;
            /** Num Days */
            num_days: number;
            /** Target Work Min */
            target_work_min: number;
            /** Demand Ids */
            demand_ids: number[];
            /** Previous Roster Id */
            previous_roster_id?: number | null;
        };
        /** RosterOut */
        RosterOut: {
            /** Id */
            id: number;
            /** Profile Id */
            profile_id: number;
            /** Name */
            name: string;
            status: components["schemas"]["RosterStatus"];
            /**
             * Roster Start
             * Format: date
             */
            roster_start: string;
            /** Num Days */
            num_days: number;
            /** Target Work Min */
            target_work_min: number;
            /** Celery Task Id */
            celery_task_id: string | null;
            /** Result */
            result: Record<string, never> | null;
        };
        /**
         * RosterStatus
         * @enum {string}
         */
        RosterStatus: "running" | "draft" | "approved" | "failed";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}